use and test dedicated ddt operator
