+++ test/dsl/timeIntegration/timeIntegration.cpp
@@ -11,41 +11,7 @@
 #include "NeoFOAM/core/dictionary.hpp"
 #include "NeoFOAM/core/parallelAlgorithms.hpp"
 #include "NeoFOAM/dsl/equation.hpp"
-#include "NeoFOAM/dsl/timeIntegration/timeIntegration.hpp"
-
-class TimeOperator : public OperatorMixin
-{
-
-public:
-
-    TimeOperator(const Executor& exec, VolumeField& field) : OperatorMixin(exec), field_(field) {}
-
-    std::string getName() const { return "TimeOperator"; }
-
-    Operator::Type getType() const { return Operator::Type::Temporal; }
-
-    void explicitOperation(NeoFOAM::Field<NeoFOAM::scalar>& source, NeoFOAM::scalar scale)
-    {
-        auto sourceField = source.span();
-        NeoFOAM::parallelFor(
-            source.exec(),
-            {0, source.size()},
-            KOKKOS_LAMBDA(const size_t i) { sourceField[i] += 1 * scale; }
-        );
-    }
-
-    VolumeField& volumeField() const { return field_; }
-
-    size_t getSize() const { return field_.internalField().size(); }
-
-    const Executor& exec() const { return exec_; }
-
-    const Executor exec_;
-
-    std::size_t nCells_;
-
-    VolumeField& field_;
-};
+#include "NeoFOAM/dsl/timeIntegration/ddt.hpp"
 
 
 TEST_CASE("TimeIntegration")
@@ -75,8 +41,7 @@ TEST_CASE("TimeIntegration")
 
     SECTION("Create equation and perform explicitOperation on " + execName)
     {
-
-        Operator ddtOperator = TimeOperator(exec, vf);
+        Operator ddtOperator = NeoFOAM::dsl::temporal::Ddt(exec, vf);
 
         auto eqn = ddtOperator + dummy;
